requirements.txt
Django==4.2.3
djangorestframework==3.14.0
drf-yasg==1.21.10
django-cors-headers==4.7.0

psycopg2-binary==2.9.10
gunicorn==23.0.0
whitenoise==6.9.0
python-dotenv==1.1.1
dj-database-url==3.0.1
